+++ xarray/core/concat.py
@@ -422,6 +422,13 @@ def _dataset_concat(
     """
     from .dataset import Dataset
 
+    datasets = list(datasets)
+
+    if not all(isinstance(dataset, Dataset) for dataset in datasets):
+        raise TypeError(
+            "The elements in the input list need to be either all 'Dataset's or all 'DataArray's"
+        )
+
     dim, coord = _calc_concat_dim_coord(dim)
     # Make sure we're working on a copy (we'll be loading variables)
     datasets = [ds.copy() for ds in datasets]
@@ -541,7 +548,14 @@ def _dataarray_concat(
     join: str = "outer",
     combine_attrs: str = "override",
 ) -> "DataArray":
+    from .dataarray import DataArray
+
     arrays = list(arrays)
+
+    if not all(isinstance(array, DataArray) for array in arrays):
+        raise TypeError(
+            "The elements in the input list need to be either all 'Dataset's or all 'DataArray's"
+        )
 
     if data_vars != "all":
         raise ValueError(